feat: Add support for NSO controllers
This commit adds support to the `hid-nintendo` driver for the Nintendo
Switch Online controllers for the NES, SNES, and Genesis. These changes
were first implemented in:

* https://github.com/DanielOgorchock/linux/pull/35

The changes have been adapted here for the latest version of the
`hid-nintendo` driver as of Linux kernel version 5.16.0 (see previous
commit).

For more about the controllers, see:

* https://www.nintendo.com/store/products/nintendo-entertainment-system-controllers/
* https://www.nintendo.com/store/products/super-nintendo-entertainment-system-controller/
* https://www.nintendo.com/store/products/sega-genesis-control-pad/

Much of these changes make reference to work by Nadia Holmquist
Pedersen, adapted from these commits and ported forward to the latest
version of the `hid-nintendo` driver:

* https://github.com/nadiaholmquist/linux/commit/21a824e
* https://github.com/nadiaholmquist/linux/commit/bdf4626

In addition, I implemented Sega Genesis controller support. At this
point, Bluetooth support for this controller is preliminary and
incomplete.

I have relicensed these changes under the GPL-2.0 in this repository,
using the same licenses as the original.

Signed-off-by: Emily Strickland <linux@emily.st>