Use reported type to distinguish controllers
To distinguish what type of controller we're dealing with, I've changed
to relying upon the `ctlr->ctlr_type` in all possible scenarios. This
appears generally to be more reliable (given especailly how the Genesis
controller seems to work).

However, I did not do this in the case of the Pro and the charging grip.
Apparently the charging grip reports itself as a Pro controller, but it
has its own product ID. Therefore, for these two controllers, I have to
look at the product ID still.

(Of course, I'd like to confirm this with a real device, but at present,
I do not have one, so that is not possible.)

I've also converted all the macros to static inline methods for
readability's sake. Hopefully GCC will compile to equivalent assembly.