Rename driver to `hid-nx` -> v1.11
This commit renames the driver from `hid-nintendo` to `hid-nx`. This was
done to avoid confusion with the default `hid-nintendo` driver.

Thanks to how DKMS and the kernel Makefile work, the resulting
`modules.alias` will tell the kernel to use this driver for all Nintendo
Switch controllers, rather than the `hid-nintendo` driver, making
further configuration unnecessary (hopefully).

Version was bumped to v1.11 because of the rename.

Several bits of documentation were modified.