Define LED_FUNCTION_PLAYER* if not defined (#1)
This change adds some preprocessor defines that were added to the Linux
kernel in 5.16. We simply define them if they're not already defined.
Even if this somehow slightly changes compatibility somehow, it's
harmless. It just affects a device name. Maybe someone's udev rule would
break. But the chances are extremely remote for reasons I don't want to
belabor here.

Fixes #1.